CB-10390: ADB doesn't recognize existing devices in Medic
In Apache CI, we get the error: ADB doesn't recognize existing devices
in Medic. The root cause for this error is that there is an additional
space in the pattern. As a result ‘emulator-5554’ will not be matched
with ‘emulator ‘ pattern (due to additional space). On removing the
additional space, the pattern matching working fine.

Also, when you run adb devices command, there will be a header line
“List of devices” will be displayed which must be ignored. Added the
code to ignore that also.